requirements.txt
-f /home/matt/.pip/wheels
PyYAML==3.11
argparse==1.2.1
arrow==0.4.4
certifi==14.05.14
future==0.13.0
memoized-property==1.0.2
mendeley==0.1.1
oauthlib==0.6.3
python-dateutil==2.2
requests==2.4.0
requests-oauthlib==0.4.1
six==1.7.3
wsgiref==0.1.2
